fix naming for protos
