+++ src/Proto.Actor/Actor.cs
@@ -10,6 +10,6 @@ namespace Proto
 {
     public static class Actor
     {
-        public static readonly Task Done = Task.FromResult(0);
+        public static readonly Task Done = Task.CompletedTask;
     }
 }
