+++ src/Proto.Actor/Proto.Actor.csproj
@@ -17,4 +17,8 @@
   <ItemGroup>
     <Protobuf Include="Protos.proto" GrpcServices="Server" />
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Utils" />
+  </ItemGroup>
 </Project>
